requirements.txt
stripe
django
django-environ